Move Config class definition into itopapi instead of model
