+++ .gitignore
@@ -1,0 +1,3 @@
+.idea
+*.pyc
+itop-cli.cfg
