+++ itop-cli
@@ -8,8 +8,9 @@ itop-cli is a basic CLI interface for iTop REST interface
 __version__ = '1.0'
 __authors__ = ['Guillaume Philippon <guillaume.philippon@lal.in2p3.fr>']
 
-from itopapi.model import ItopapiRack, ItopapiServer, Config
+from itopapi.model import ItopapiRack, ItopapiServer
 from itopapi.view import QuattorJson
+from itopapi.config import Config
 import argparse
 import ConfigParser
 import json
