+++ itopapi/__init__.py
@@ -8,3 +8,4 @@ __version__ = '1.0'
 __authors__ = ['Guillaume Philippon <guillaume.philippon@lal.in2p3.fr>']
 
 from itopapi.model import ItopapiRack, ItopapiServer
+from itopapi.config import Config
